Move GameWorld into its own file
